% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemaps.R
\name{addCartoBasemap}
\alias{addCartoBasemap}
\title{Add an authenticated CARTO basemap}
\usage{
addCartoBasemap(
  map,
  style = c("positron", "dark_matter", "voyager"),
  api_key = Sys.getenv("CARTO_API_KEY"),
  group = NULL,
  layerId = NULL,
  mode = c("vector", "raster"),
  opacity = 1,
  timeout = 30
)
}
\description{
Add an authenticated CARTO basemap
}
