% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemaps.R
\name{addIowaBasemap}
\alias{addIowaBasemap}
\title{Add an Iowa DOT vector basemap}
\usage{
addIowaBasemap(
  map,
  style = c("greyscale", "dark", "color"),
  group = NULL,
  layerId = NULL,
  labels = TRUE,
  item_id = NULL,
  source = c("live", "bundled"),
  fallback = FALSE,
  refresh = FALSE,
  opacity = 1,
  timeout = 30
)
}
\description{
Add an Iowa DOT vector basemap
}
