% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{checkDOTBasemaps}
\alias{checkDOTBasemaps}
\title{Check Web Map and child-layer modification dates and definitions}
\usage{
checkDOTBasemaps(timeout = 30)
}
\description{
Check Web Map and child-layer modification dates and definitions
}
